Edited errors and organized files for testing
Bridging.java
Changed class access modifier to public
Changed test expected result
