+++ src/engine/src/rabbitescape/engine/WorldChanges.java
@@ -9,6 +9,7 @@ import rabbitescape.engine.World.NoBlockFound;
 import rabbitescape.engine.World.NoSuchAbilityInThisWorld;
 import rabbitescape.engine.World.NoneOfThisAbilityLeft;
 import rabbitescape.engine.World.UnableToAddToken;
+import rabbitescape.engine.tokens.TokenFactory;
 import rabbitescape.engine.util.Position;
 
 public class WorldChanges
