+++ src/engine/src/rabbitescape/engine/solution/SandboxGame.java
@@ -13,10 +13,10 @@ import rabbitescape.engine.Pipe;
 import rabbitescape.engine.Rabbit;
 import rabbitescape.engine.Thing;
 import rabbitescape.engine.Token;
-import rabbitescape.engine.TokenFactory;
 import rabbitescape.engine.VoidMarkerStyle;
 import rabbitescape.engine.World;
 import rabbitescape.engine.textworld.Comment;
+import rabbitescape.engine.tokens.TokenFactory;
 
 /**
  * A completely sandboxed game that can be edited and have solutions run against
