+++ src/engine/src/rabbitescape/engine/textworld/LineProcessor.java
@@ -28,8 +28,8 @@ import rabbitescape.engine.Pipe;
 import rabbitescape.engine.Rabbit;
 import rabbitescape.engine.Thing;
 import rabbitescape.engine.Token;
-import rabbitescape.engine.TokenFactory;
 import rabbitescape.engine.VoidMarkerStyle;
+import rabbitescape.engine.tokens.TokenFactory;
 import rabbitescape.engine.util.Dimension;
 import rabbitescape.engine.util.MegaCoder;
 import rabbitescape.engine.util.Position;
