+++ src/engine/src/rabbitescape/engine/tokens/BashToken.java
@@ -1,4 +1,4 @@
-package rabbitescape.engine;
+package rabbitescape.engine.tokens;
 
 import static rabbitescape.engine.ChangeDescription.State.TOKEN_BASH_FALLING;
 import static rabbitescape.engine.ChangeDescription.State.TOKEN_BASH_FALL_TO_SLOPE;
@@ -6,6 +6,8 @@ import static rabbitescape.engine.ChangeDescription.State.TOKEN_BASH_ON_SLOPE;
 import static rabbitescape.engine.ChangeDescription.State.TOKEN_BASH_STILL;
 
 import rabbitescape.engine.ChangeDescription.State;
+import rabbitescape.engine.Token;
+import rabbitescape.engine.World;
 
 public class BashToken extends Token {
 	public BashToken(int x, int y) {
